chenge store folder structure, move filter action and reducers to own folder
