+++ src/store/filter/actions/filter.action.ts
@@ -1,5 +1,5 @@
 import { createCustomAction } from 'typesafe-actions';
-import { Data } from '../../pages/comparison/interfaces/filter.interface';
+import { Data } from '../../../pages/comparison/interfaces/filter.interface';
 
 export const filterAction = createCustomAction('FILTER', (data: Data | null) => {
   return { payload: data };
