+++ src/store/root.reducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { resultReducer } from './result/reducers/result.reducer';
-import filterReducer from './filter.reducer';
+import filterReducer from './filter/reducers/filter.reducer';
 import { petReducer } from './reducers/pet.reducer';
 import comparisonReducer from './reducers/comparison.reducer';
